chore: refactor kommun to own class
